refactor: extract class DataUtils
