+++ atividade-1.2/src/models/PacienteModel.js
@@ -1,6 +1,6 @@
 const PacientesTable = require('../database/PacientesTable');
 const CPFValidator = require('../validators/CPFValidator');
-// const PacienteValidator = require('../validators/PacienteValidator');
+const DataValidator = require('../validators/DataValidator');
 
 module.exports = class PacienteModel {
   constructor(body) {
@@ -23,21 +23,7 @@ module.exports = class PacienteModel {
   }
 
   validaData() {
-    const dataReg = /^(\d{2})\/(\d{2})\/(\d{4})$/;
-    const match = this.body.dataNascimento.match(dataReg);
-    if (!match) {
-      this.errors.push('data nascimento não está no formato DD/MM/AAAA');
-      return;
-    }
-    const dia = parseInt(match[1], 10);
-    const mes = parseInt(match[2], 10);
-    const ano = parseInt(match[3], 10);
-
-    const data = new Date(ano, mes - 1, dia);
-    const dataInvalida = data.getFullYear() !== ano
-      || data.getMonth() + 1 !== mes
-      || data.getDate() !== dia;
-    if (dataInvalida) this.errors.push('data nascimento inválida');
+    if (!DataValidator.valida(this.body.dataNascimento)) this.errors.push('data nascimento inválida');
   }
 
   validaCpf() {
